First commit
This lays out the main idea of this tool, which is to sort pictures
automatically. For the time being, the logic we follow is straight forward:
place the pictures in folders following the convention "year/month/day".

Ideally, this will be customizable. Also, it depends on the pictures, as
it won't work if they don't have EXIF data.