Port to C++
src.pro: defines what files are goind to be taken into account by 'qmake'.

TO DO: I need 'something' to make the whole process of compilation even
easier because this Makefile is actually autogenerated. Or do I prefer to
keep it and manually change it when new files come into place?