src/ui: Add user interface using wx library
wx offers portability across platforms which makes it ideal if I want to
run my app on Windows machines. I also considered Qt4, but compiling
binary files for Windows was extremely complicated, so I decided to go
for wx instead.

In order to compile different binaries, I need to have machines with the
target OS. In the future, I'd like to explore what are the possibilities
to cross-compile the different versions from Linux.